change the default pick symbol
