+++ src/javascript/app/Stores/Modules/Trading/Helpers/active-symbols.js
@@ -1,13 +1,9 @@
 import { flow }            from 'mobx';
-import { WS }              from 'Services';
-import { localize }        from '_common/localize';
-import { redirectToLogin } from '_common/base/login';
 import BinarySocket        from '_common/base/socket_base';
-
-export const pickDefaultSymbol = (active_symbols = []) => {
-    if (!active_symbols.length) return '';
-    return active_symbols.filter(symbol_info => /major_pairs|random_index/.test(symbol_info.submarket))[0].symbol;
-};
+import { localize }        from '_common/localize';
+import { LocalStore }      from '_common/storage';
+import { redirectToLogin } from '_common/base/login';
+import { WS }              from 'Services';
 
 export const showUnavailableLocationError = flow(function* (showError) {
     const website_status = yield BinarySocket.wait('website_status');
@@ -34,3 +30,36 @@ export const isMarketClosed = (active_symbols = [], symbol) => {
         :
         false;
 };
+
+export const pickDefaultSymbol = (active_symbols = []) => {
+    if (!active_symbols.length) return '';
+    return getFavoriteOpenSymbol(active_symbols) || getFirstOpenSymbol(active_symbols);
+};
+
+const getFavoriteOpenSymbol = (active_symbols) => {
+    const chart_favorites = LocalStore.get('cq-favorites');
+    if (!chart_favorites) return undefined;
+    const client_favorite_markets = JSON.parse(chart_favorites)['chartTitle&Comparison'];
+
+    const client_favorite_list = client_favorite_markets
+        .map(client_fav_symbol => active_symbols
+            .find(symbol_info => symbol_info.symbol === client_fav_symbol));
+    if (client_favorite_list) {
+        const client_first_open_symbol = client_favorite_list
+            .filter(symbol => symbol).find(isSymbolOpen);
+        if (client_first_open_symbol) return client_first_open_symbol.symbol;
+    }
+    return undefined;
+};
+
+const getFirstOpenSymbol = (active_symbols) => {
+    const first_open_symbol = active_symbols
+        .filter(symbol_info => /major_pairs|random_index/.test(symbol_info.submarket))
+        .find(isSymbolOpen);
+    if (first_open_symbol) return first_open_symbol.symbol;
+    return active_symbols.find(symbol_info => symbol_info.submarket === 'major_pairs').symbol;
+};
+
+const isSymbolOpen = (symbol) => (
+    symbol.exchange_is_open === 1
+);
